Address feedback for Q8
Q8
Reviewer Comments
There is nothing wrong with your answer per se. However, this question is about PCA v.s. ICA. Would you be able to adjust your answer to reflect this in your next submission?
I understand that the working of the question is vague - we will change that soon!