Update to assign state to self.state for project specs;
Reviewer Comments
The agent does not update it's state while running. This is a minor problem, and easily fixed. I've included a note in the code for it. You should be able to see the agents state on the pygame display. In the attached screenshot, you can see that state: None. This is the case for the entire run, and does not comply with the spec.